requirements.txt
# Requirements for the heartbeat project
# Minimum recommended versions — adjust if you have known constraints.
pandas>=1.3
numpy>=1.21
scikit-learn>=1.0
kafka-python>=2.0.2
tensorflow>=2.9
requests>=2.25
dash>=2.0
plotly>=5.0

# Optional / environment notes:
# - If you only need model serving (not training), you can use tensorflow-cpu to avoid GPU drivers.
# - Kafka broker and TensorFlow Serving are external services; see README or notes below for Docker commands.
Flask>=2.0
flask-cors>=3.0